% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_CombiningPValuesFinal.r
\name{CombinationPValues}
\alias{CombinationPValues}
\title{CombinationPValues}
\usage{
myClass(x)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
Functions within this class allow for multiple p values

n p-values used as input

Combination p-value using Fishers method

Combination p-value using Pearson method

Combination p-value using George method

Combination p-value using Edgington method

Combination p-value using Stouffer method

Combination p-value using Tippett method
}
\description{
Functions within this class allow for multiple p values
to be defined within each method below
Method options include: Fisher, Pearson, Edgington, Stouffer, George, and
Tippett
}
